ai-ops/requirements.txt
requests
pyyaml
kubernetes
openai
anthropic